Remove unnecessary boxing and unboxing.
git-svn-id: https://svn.apache.org/repos/asf/tika/trunk@1612701 13f79535-47bb-0310-9956-ffa450edef68

+++ tika-parsers/src/main/java/org/apache/tika/parser/microsoft/SummaryExtractor.java
@@ -192,7 +192,7 @@ public class SummaryExtractor {
                     metadata.set(prop, ((Long)value).intValue());
                 } else if (value instanceof Double) {
                     Property prop = Property.externalReal(key);
-                    metadata.set(prop, ((Double)value).doubleValue());
+                    metadata.set(prop, (Double)value);
                 } else if (value instanceof Integer) {
                     Property prop = Property.externalInteger(key);
                     metadata.set(prop, ((Integer)value).intValue());
